# case-0.1.0-client/src/main/java/com/afkl/travel/exercise/clientkit/model/LocationType.java
/*
 * Travel API
 * You can use this api to retrieve locations.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: someone@airfranceklm.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package com.afkl.travel.exercise.clientkit.model;

import java.util.Objects;
import java.util.Arrays;
import com.google.gson.annotations.SerializedName;

import java.io.IOException;
import com.google.gson.TypeAdapter;
import com.google.gson.annotations.JsonAdapter;
import com.google.gson.stream.JsonReader;
import com.google.gson.stream.JsonWriter;

/**
 * Gets or Sets LocationType
 */
@JsonAdapter(LocationType.Adapter.class)
public enum LocationType {
  
  COUNTRY("country"),
  
  CITY("city"),
  
  AIRPORT("airport");

  private String value;

  LocationType(String value) {
    this.value = value;
  }

  public String getValue() {
    return value;
  }

  @Override
  public String toString() {
    return String.valueOf(value);
  }

  public static LocationType fromValue(String value) {
    for (LocationType b : LocationType.values()) {
      if (b.value.equals(value)) {
        return b;
      }
    }
    throw new IllegalArgumentException("Unexpected value '" + value + "'");
  }

  public static class Adapter extends TypeAdapter<LocationType> {
    @Override
    public void write(final JsonWriter jsonWriter, final LocationType enumeration) throws IOException {
      jsonWriter.value(enumeration.getValue());
    }

    @Override
    public LocationType read(final JsonReader jsonReader) throws IOException {
      String value = jsonReader.nextString();
      return LocationType.fromValue(value);
    }
  }
}